[apps/interactive_curve_view_range] Check preferences for autozoom memoization, to reset autozoom when angle unit or complex format are changed

+++ apps/regression/graph/graph_controller.h
@@ -42,8 +42,8 @@ class GraphController : public Shared::InteractiveCurveViewController {
   bool moveCursorVertically(OMG::VerticalDirection direction) override;
 
   // InteractiveCurveViewRangeDelegate
-  uint32_t autoZoomChecksum() const override {
-    return m_store->storeChecksum();
+  uint64_t autoZoomChecksum() const override {
+    return static_cast<uint64_t>(m_store->storeChecksum());
   }
   Poincare::Range2D optimalRange(
       bool computeX, bool computeY,
